message handlers wip inc. permissions events

+++ src/SFA.DAS.ProviderRelationships/Models/Permission.cs
@@ -1,4 +1,5 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
+using SFA.DAS.EmployerCommitments.Messages.Events;
 
 namespace SFA.DAS.ProviderRelationships.Models
 {
@@ -7,7 +8,7 @@ namespace SFA.DAS.ProviderRelationships.Models
         public virtual int PermissionId { get; set; }
         public virtual PermissionType Type { get; set; }
 
-        public virtual int AccountLegalEntityId { get; set; }
+        public virtual long AccountLegalEntityId { get; set; }
 
         //[ForeignKey("AccountLegalEntityId")]
         public virtual AccountLegalEntity AccountLegalEntity { get; set; }
